Continuing the carbonapi and zipper code merge. Removed multiple obsolete zipper metrics. Removed metrics namespace for zipper. Removed obsolete zipper logger.
