+++ pkg/app/zipper/app.go
@@ -7,8 +7,6 @@ import (
 	"time"
 
 	bnet "github.com/bookingcom/carbonapi/pkg/backend/net"
-	"github.com/pkg/errors"
-	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/dgryski/go-expirecache"
 
@@ -28,15 +26,6 @@ type App struct {
 
 	Metrics *PrometheusMetrics
 	Lg      *zap.Logger
-}
-
-// Start start launches the goroutines starts the app execution
-// `server` and `promPortOverride` are temporary to implement api and zipper merge.
-// TODO: Clean-up this function after merge is done.
-func (app *App) Start(lg *zap.Logger) {
-	go probeTopLevelDomains(app.TopLevelDomainCache, app.TLDPrefixes, app.Backends, app.Config.InternalRoutingCache, app.Metrics)
-
-	metricsServer(app)
 }
 
 // InitBackends inits backends.
@@ -62,10 +51,6 @@ func InitBackends(config cfg.Zipper, ms *PrometheusMetrics, logger *zap.Logger) 
 		var b backend.Backend
 		var err error
 
-		limiterExits, err := ms.BackendLimiterExits.CurryWith(prometheus.Labels{"backend": host.Http})
-		if err != nil {
-			return nil, errors.Wrap(err, "currying the metric for LimiterExits failed")
-		}
 		bConf := bnet.Config{
 			Address:            host.Http,
 			DC:                 dc,
@@ -76,10 +61,6 @@ func InitBackends(config cfg.Zipper, ms *PrometheusMetrics, logger *zap.Logger) 
 			QHist:              ms.TimeInQueueSeconds,
 			Responses:          ms.BackendResponses,
 			Logger:             logger,
-			ActiveRequests:     ms.ActiveUpstreamRequests.WithLabelValues(host.Http),
-			WaitingRequests:    ms.WaitingUpstreamRequests.WithLabelValues(host.Http),
-			LimiterEnters:      ms.BackendLimiterEnters.WithLabelValues(host.Http),
-			LimiterExits:       limiterExits,
 		}
 		var be backend.BackendImpl
 		if host.Grpc != "" {
